+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Seed phrase generator
+- Seed phrase validator
+- Service: generate seed phrase
+- Extended private key serialization
+- Extended public key serialization
+- Private key exportation
+- Private key importation
+- Key derivation algorithm implementation
 
 ## [0.1.0] - 2018-06-18
 ### Added